renamed browser compatible file
