+++ gas-price-notifier/assets/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "semantic.min.js",
   "scripts": {
-    "build": "browserify ./worker.js -o ./worker.build.js",
+    "clean": "rm worker.js",
+    "build": "browserify ./_worker.js -o ./worker.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
